requirements.txt
notion-client
python-dotenv
requests
pandas
matplotlib
seaborn
plotly.express
kaleido